kallsyms: use ebpf alternative when kallsyms not available
On the minikube kernel 5.10.57, /proc/kallsyms does not provide the
following symbols:
- bpf_prog_fops (used by the top-ebpf gadget)
- socket_file_ops (used by the socket enricher)

This patch provides an alternative to get the address of those symbols
using ebpf.

Most of the required ebpf code was already implemented for the
container-hook package. The file is moved to make it available in a
generic package:
pkg/container-hook/bpf/privatedata.bpf.c -> pkg/kfilefields/bpf/filefields.bpf.c

Co-authored-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>
Co-authored-by: Alban Crequy <albancrequy@linux.microsoft.com>
Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>