wasm: Use cache to speed-up things
Use cached compiled wasm module for GetGadgetInfo/RunGadget
to speed up things. The cache is using wazero.CompilationCache,
which is in-memory caching so we implement some logic to ensure
cache is cleaned/refreshed to avoid memory hit. Also, to explain
more why we need to refresh cache is host modules are always
added to cache with different ID so it is bit tricky to clean them.
The main (gadget) module is deleted once closed so caching is
only valid for single gadget run.

Signed-off-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>